requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
timm
pandas
pillow
scikit-learn
seaborn
matplotlib
supabase
huggingface-hub